Update catdateweight.csv
For update 20hrs ago we found (by the weight change being anomalous, and wondering why) the collar weight I was using was wrong: I was using 15g, but should have been 32g - I weighed the collar again tonight to be sure, and get 33g (slight imbalance across the scale I think). I'm going with a last revision - to adjust for 33g instead of 32g, and calling it set. There was also a 1g tie that was used on some collar-wearing days, but am not adjusting for that. Then also just now I saw that I had yesterday's date one-day-off, so correcting for that. Should be all back on track now.